Solve issues w/ 'simplifyRegularExpression(..)': Allow passing regexp directly (Comparer, ..) and hence skip production of regexp
In jogamp, we like to exclude 'jogamp/**' packages,
however 'simplifyRegularExpression(..)'
allows this to match 'com/jogamp/lala/Lala'.

This is enforced by tests in 'DifferenceAccumulatingHandlerTest'
and hence we have to consider this to be desired !? (IMHO this is a bug).

However, to not break backward capability, this patch allows passing
a flag for the includes and excludes whether they are regexp already
and hence we can skip 'simplifyRegularExpression(..)' and avoid this dilema.